Create thin archive and convert it afterwards
While this worked fine on rawhide, the MRI script is not accepted
on f37. Apparently + is a special character and something we end
up trying to use libc rather than libc++.a.

Avoid this by creating a thin archive first and then only using
the MRI script to convert it into a fat archive. Weird that there
is apparently no native way to do this.